Save and restore pane position, better handle missing keys in config file
Previously when a key was missing in the config file, we stopped further processing of the file and used default values for all keys, which could result in a loss of information.
Now we use default values only for the missing keys and process them all without losing existing data.